building telegram bot

+++ csd_telebot.py
@@ -4,26 +4,25 @@ from telegram.ext import Updater, MessageHandler, Filters, CommandHandler, Conve
 from telegram import ReplyKeyboardMarkup, KeyboardButton
 
 from csdbot_token import TOKEN
-from launch import launch
-
-
-# SCRIPT GLOBALS
-NORMAL_MODE = False
-PARAMS_UPLOADED = False
-LAUNCH_TEXT = 'ПАРАМЕТРЫ ОПЕРАЦИИ:\n'
-FNAME = None
-EXTENSION = None
+import launch
 
 # USER_DATA FIELDS
+MODE_CHOICE = 'mode_choice'
 TASK = 'task'
 EXT = 'ext'
 SPAN = 'span'
-MODE = 'mode'
+DEMO = 'demo'
 UPLOAD = 'upload'
 LAUNCH = 'launch'
+PARAMS_UPLOADED = 'params_uploaded'
+LAUNCH_TEXT = 'launch_text'
+FNAME = 'fname'
+OUTF = 'outf'
+EXTENSION = 'extension'
+PARAMS_SOURCE = 'params_source'
 
 # Значения клавиатуры
-DEMO = 'Демо'
+DEMO_MODE = 'Демо'
 NORMAL = 'Обычный'
 INFO = 'ИНФО'
 CONTR = 'По контрактам'
@@ -35,7 +34,7 @@ START = 'Запуск'
 CANCEL = 'Снять задачу'
 
 
-keyboard_mode = [[KeyboardButton(text=DEMO),
+keyboard_mode = [[KeyboardButton(text=DEMO_MODE),
              KeyboardButton(text=NORMAL)]]
 markup_mode = ReplyKeyboardMarkup(keyboard_mode, one_time_keyboard=True)
 
@@ -49,21 +48,20 @@ keyboard_ext = [[KeyboardButton(text=CSV),
              KeyboardButton(text=JSON)]]
 markup_ext = ReplyKeyboardMarkup(keyboard_ext, one_time_keyboard=True)
 
-keyboard_launch = [[KeyboardButton(text='Запуск'),
-                KeyboardButton(text='Снять задачу')]]
-markup_launch = ReplyKeyboardMarkup(keyboard_ext, one_time_keyboard=True)
+keyboard_launch = [[KeyboardButton(text=START),
+                KeyboardButton(text=CANCEL)]]
+markup_launch = ReplyKeyboardMarkup(keyboard_launch, one_time_keyboard=True)
 
 
-DIALOG_MAPPING = {MODE: 0,
+DIALOGUE = {MODE_CHOICE: 0,
                 UPLOAD: 1,
                 TASK: 2,
                 EXT: 3,
-                SPAN: 4,
-                LAUNCH: 5}
+                SPAN: 4}
 
 
 def is_demo(choice):
-    return choice == DEMO
+    return choice == DEMO_MODE
 
 
 def get_task_code(choice):
@@ -100,20 +98,20 @@ def get_fextension(choice):
 
 def start(bot, update):
     update.message.reply_text("Выберите режим", reply_markup=markup_mode)
-    return 0
+    return DIALOGUE[MODE_CHOICE]
 
 
 def mode(bot, update, user_data):
     mode = update.message.text
-    LAUNCH_TEXT += 'Режим: {}\n'.format(mode)
-    print(mode, is_normal_mode(mode))
-    user_data['demo'] = is_demo(mode)
+    user_data[DEMO] = is_demo(mode)
+    user_data[LAUNCH_TEXT] = 'ПАРАМЕТРЫ ОПЕРАЦИИ:\n'
+    user_data[LAUNCH_TEXT] += 'Режим: {}\n'.format(mode)
+    print(user_data[LAUNCH_TEXT])
     if is_demo(mode):
         update.message.reply_text("Выберите задачу", reply_markup=markup_task)
-        return 2
-    print('Normal mode is', NORMAL_MODE)
-    NORMAL_MODE = True
-    return 1
+        print(DIALOGUE[TASK])
+        return DIALOGUE[TASK]
+    return DIALOGUE[UPLOAD]
 
 
 def upload(bot, update, user_data):
@@ -122,69 +120,74 @@ def upload(bot, update, user_data):
         received_file = bot.getFile(update.message.document.file_id)
         if received_file:
             PARAMS_UPLOADED = True
-        received_file.download('params/params_csv.csv')
-        return 1
+        received_file.download(os.path.join('params', 'params_telebot.csv'))
+        return DIALOGUE[UPLOAD]
     update.message.reply_text("Выберите задачу", reply_markup=markup_task)
-    return 2
+    user_data[PARAMS_SOURCE] = os.path.join('params', 'params_telebot.csv')
+    return DIALOGUE[TASK]
 
 
 def task(bot, update, user_data):
     task = update.message.text
-    user_data['task'] = get_task_code(task)
+    user_data[TASK] = get_task_code(task)
     print(task)
     if task == PROD or task == CONTR:
-        FNAME = '{}'.format(update.message.chat_id)
-        LAUNCH_TEXT += 'Задача: выгрузка "{}"\n'.format(task)
+        user_data[FNAME] = '{}'.format(update.message.chat_id)
+        user_data[LAUNCH_TEXT] += 'Задача: выгрузка "{}"\n'.format(task)
+        print(user_data[LAUNCH_TEXT])
         update.message.reply_text("Выберите формат файла", reply_markup=markup_ext)
-        return 3
-    else:
-        LAUNCH_TEXT += 'Задача: Вывести информацию о запросе\n'
-        update.message.reply_text("Вы можете указать число дней в подпериодах, на которые может быть разбит период для дробления запроса. Число дней указывается целым числом. Если вы не хотите указывать этот параметр, просто нажмите кнопку 'Запуск'", reply_markup=markup_launch)
-        return 4
+        return DIALOGUE[EXT]
+    user_data[LAUNCH_TEXT] += 'Задача: Вывести информацию о запросе\n'
+    print(user_data[LAUNCH_TEXT])
+    update.message.reply_text('Вы можете указать число дней в подпериодах, на которые может быть разбит период для дробления запроса. Число дней указывается целым числом. Если вы не хотите указывать этот параметр, просто нажмите кнопку "Запуск"', reply_markup=markup_launch)
+    return DIALOGUE[SPAN]
 
 
 def outformat(bot, update, user_data):
     outf = update.message.text
-    user_data['format'] = outf
+    user_data[OUTF] = outf
     print(outf)
-    LAUNCH_TEXT += 'Формат выгрузки: {}\n'.format(outf)
-    EXTENSION = get_fextension(outf)
+    user_data[LAUNCH_TEXT] += 'Формат выгрузки: {}\n'.format(outf)
+    user_data[EXTENSION] = get_fextension(outf)
     update.message.reply_text("Вы можете указать число дней в подпериодах, на которые может быть разбит период для дробления запроса. Число дней указывается целым числом. Если вы не хотите указывать этот параметр, просто нажмите кнопку 'Запуск'", reply_markup=markup_launch)
-    return 4
+    return DIALOGUE[SPAN]
 
 
 def add_span(bot, update, user_data):
     choice = update.message.text
     if choice == START:
         launch_launch(bot, update, user_data)
+        return ConversationHandler.END
     elif choice == CANCEL:
         update.message.reply_text('Задача снята.')
+        if not user_data[DEMO]:
+            os.remove(os.path.join('params', 'params_telebot.csv'))
         return ConversationHandler.END
     span = process_span(choice)
     if type(span) is str:
         update.message.reply_text(span)
-        return 4
+        return DIALOGUE[SPAN]
     user_data['span'] = span
-    LAUNCH_TEXT += 'Длина подпериодов (в днях): {}\n'.format(span)
+    user_data[LAUNCH_TEXT] += 'Длина подпериодов (в днях): {}\n'.format(span)
     launch_launch(bot, update, user_data)
 
 
 def launch_launch(bot, update, user_data):
-    LAUNCH_TEXT += 'Приступаю к выполнению. В зависимости от размера выгрузки операция может занять некоторое время.'
-    result = launch.launch(source='params/params_csv.csv',
-                    task=user_data['task'],
-                    out_format=user_data.get('format', 'CSV'),
-                    span=user_data.get('span', 30),
-                    out_name=FNAME,
-                    demo=user_data['demo'])
     chat_id = update.message.chat_id
-    f_path = os.path.join('data', FNAME + EXTENSION)
-    bot.send_document(chat_id=chat_id, document=open(f_path, 'r'))
+    user_data[LAUNCH_TEXT] += 'Приступаю к выполнению. В зависимости от размера выгрузки операция может занять некоторое время.'
+    bot.send_message(chat_id=chat_id, text=user_data[LAUNCH_TEXT])
+    result = launch.launch(source=user_data.get(PARAMS_SOURCE),
+                    task=user_data[TASK],
+                    out_format=user_data.get(OUTF, 'CSV'),
+                    span=user_data.get(SPAN, 30),
+                    out_name=user_data.get(FNAME),
+                    demo=user_data[DEMO])
+    if user_data.get(FNAME):
+        f_path = os.path.join('data', user_data[FNAME] + user_data[EXTENSION])
+        bot.send_document(chat_id=chat_id, document=open(f_path, 'r'))
+        os.remove(f_path)
     bot.send_message(chat_id=chat_id, text=result)
-    os.remove(f_path)
     user_data.clear()
-    return ConversationHandler.END
-
 
 
 #
@@ -198,7 +201,7 @@ def main():
     start_handler = CommandHandler('start', start)
     conversation_handler = ConversationHandler(
                 entry_points=[start_handler],
-                states={0: [RegexHandler('^(ДЕМО|Обычный)$', mode, pass_user_data=True)],
+                states={0: [RegexHandler('^(Демо|Обычный)$', mode, pass_user_data=True)],
                         1: [MessageHandler(Filters.document, upload, pass_user_data=True)],
                         2: [RegexHandler('^(ИНФО|По контрактам|По продуктам)$', task, pass_user_data=True)],
                         3: [RegexHandler('^(CSV|XLSX|JSON)$', outformat, pass_user_data=True)],
